+++ lib/src/motor_flutter_ui.dart
@@ -5,9 +5,12 @@ part of 'motor_flutter_base.dart';
 /// This extension is used to provide UI related functions to the [MotorFlutter] class.
 extension MotorFlutterUI on MotorFlutter {
   /// Display a modal dialog above the current application content which walks the user through the process
-  /// of registering a new account. An optional [onError] callback can be used to handle errors that occur
-  /// during the registration process. Returns [CreateAccountResponse] if the user successfully creates an
+  /// of registering a new account. Returns [CreateAccountResponse] if the user successfully creates an
   /// account, otherwise returns null.
+  ///
+  /// ### Parameters
+  /// - Callback [onError] can be used to handle errors that occur (optional)
+  /// - Callback [onKeysGenerated] is used to handle the dsc, and psk keys that are generated (optional)
   ///
   /// ### Example
   ///
@@ -25,16 +28,24 @@ extension MotorFlutterUI on MotorFlutter {
   ///         Get.offAll(() => const HomePage());
   ///     }
   ///   },
+  ///   onKeysGenerated: (dsc, psk) {
+  ///     print("DSC Length: ${dsc.length}"); // Output: "DSC Length: 32"
+  ///     print("PSK Length: ${psk.length}"); // Output: "PSK Length: 32"
+  ///   },
   /// );
   /// ```
-  Future<CreateAccountResponse?> showRegisterModal({ErrorCallback? onError}) async {
+  Future<CreateAccountResponse?> showRegisterModal({HandleKeysCallback? onKeysGenerated, ErrorCallback? onError}) async {
     final completer = Completer<CreateAccountResponse?>();
     if (MotorFlutter.isReady) {
       Get.dialog(RegisterModalPage(
         onCreateAccountResponse: (response) {
+          if (response == null && onKeysGenerated == null) {
+            throw Exception("onKeysGenerated callback must be provided if GetStorage is not initialized");
+          }
           completer.complete(response);
         },
         onError: onError,
+        onKeysGenerated: onKeysGenerated,
       ));
     }
     return completer.future;